+++ packages/react/transform/src/swc_plugin_compat/mod.rs
@@ -481,7 +481,9 @@ where
         call.span = pure_span;
         Expr::Call(call)
       }
-      _ => unreachable!(),
+      _ => unreachable!(
+        "Unexpected expression type in wrap_with_lynx_component_helper - expected Call expression"
+      ),
     };
 
     wrap_call
@@ -655,8 +657,12 @@ where
             span: DUMMY_SP,
           })
         }
-        JSXElementName::JSXMemberExpr(_) => unreachable!(),
-        JSXElementName::JSXNamespacedName(_) => unreachable!(),
+        JSXElementName::JSXMemberExpr(_) => {
+          unreachable!("Unexpected JSXMemberExpr in component is polyfill - expected Ident")
+        }
+        JSXElementName::JSXNamespacedName(_) => {
+          unreachable!("Unexpected JSXNamespacedName in component is polyfill - expected Ident")
+        }
       }
       if !self.opts.disable_deprecated_warning {
         HANDLER.with(|handler| {
@@ -938,7 +944,7 @@ where
         } else {
           props_key[2..].to_lowercase()
         };
-        return Some(format!("{}{}", prefix, suffix));
+        return Some(format!("{prefix}{suffix}"));
       }
       None
     }
@@ -949,11 +955,8 @@ where
           handler
             .struct_span_warn(
               n.span,
-              format!(
-                "DEPRECATED: old event props \"{}\" is changed to \"{}\"",
-                old_name, new_name
-              )
-              .as_str(),
+              format!("DEPRECATED: old event props \"{old_name}\" is changed to \"{new_name}\"",)
+                .as_str(),
             )
             .emit()
         });