+++ packages/rspeedy/plugin-react/src/entry.ts
@@ -15,7 +15,6 @@ import { LAYERS, ReactWebpackPlugin } from '@lynx-js/react-webpack-plugin'
 import type { ExposedAPI } from '@lynx-js/rspeedy'
 import { RuntimeWrapperWebpackPlugin } from '@lynx-js/runtime-wrapper-webpack-plugin'
 import {
-  CSSPlugins,
   LynxEncodePlugin,
   LynxTemplatePlugin,
   WebEncodePlugin,
@@ -190,12 +189,16 @@ export function applyEntry(
           targetSdkVersion,
 
           experimental_isLazyBundle,
-          cssPlugins: [
-            CSSPlugins.parserPlugins.removeFunctionWhiteSpace(),
-          ],
+          cssPlugins: [],
         }])
         .end()
     })
+
+    const rsbuildConfig = api.getRsbuildConfig()
+    const userConfig = api.getRsbuildConfig('original')
+
+    const enableChunkSplitting =
+      rsbuildConfig.performance?.chunkSplit?.strategy !== 'all-in-one'
 
     let finalFirstScreenSyncTiming = firstScreenSyncTiming
 
@@ -205,7 +208,8 @@ export function applyEntry(
         // TODO: support inlineScripts in lazyBundle
         inlineScripts = true
       } else {
-        inlineScripts = environment.config.output?.inlineScripts ?? true
+        inlineScripts = environment.config.output?.inlineScripts
+          ?? !enableChunkSplitting
       }
 
       if (inlineScripts !== true) {
@@ -247,14 +251,8 @@ export function applyEntry(
         .end()
     }
 
-    const rsbuildConfig = api.getRsbuildConfig()
-    const userConfig = api.getRsbuildConfig('original')
-
     let extractStr = originalExtractStr
-    if (
-      rsbuildConfig.performance?.chunkSplit?.strategy !== 'all-in-one'
-      && originalExtractStr
-    ) {
+    if (enableChunkSplitting && originalExtractStr) {
       logger.warn(
         '`extractStr` is changed to `false` because it is only supported in `all-in-one` chunkSplit strategy, please set `performance.chunkSplit.strategy` to `all-in-one` to use `extractStr.`',
       )