+++ packages/web-platform/web-worker-runtime/src/backgroundThread/background-apis/createNativeApp.ts
@@ -15,6 +15,7 @@ import {
   type BackMainThreadContextConfig,
   I18nResource,
   reportErrorEndpoint,
+  globalDisallowedVars,
 } from '@lynx-js/web-constants';
 import { createInvokeUIMethod } from './crossThreadHandlers/createInvokeUIMethod.js';
 import { registerPublicComponentEventHandler } from './crossThreadHandlers/registerPublicComponentEventHandler.js';
@@ -71,6 +72,11 @@ export async function createNativeApp(
         lynxCoreInject.tt.lynxCoreInject = lynxCoreInject;
         lynxCoreInject.tt.globalThis ??= new Proxy(lynxCoreInject, {
           get(target, prop) {
+            if (
+              typeof prop === 'string' && globalDisallowedVars.includes(prop)
+            ) {
+              return undefined;
+            }
             // @ts-expect-error
             return target[prop] ?? globalThis[prop];
           },