+++ packages/rspeedy/create-rspeedy/CHANGELOG.md
@@ -1,5 +1,7 @@
 # create-rspeedy
 
+## 0.10.7
+
 ## 0.10.6
 
 ### Patch Changes